Display correct sidebar routes on mobile and desktop
