+++ src/partials/sidebar/styles/index.js
@@ -1,5 +1,15 @@
 import SidebarSectionHeading from "./sidebar-section-heading";
 import SidebarSectionList from "./sidebar-section-subheading";
 import SidebarSectionSublist from "./sidebar-section-sublist";
+import SidebarListItem from "./sidebar-list-item";
+import SidebarListItemAnchorLink from "./sidebar-list-item-anchor-link";
+import SidebarListItemLink from "./sidebar-list-item-link";
 
-export { SidebarSectionHeading, SidebarSectionList, SidebarSectionSublist };
+export {
+  SidebarSectionHeading,
+  SidebarSectionList,
+  SidebarSectionSublist,
+  SidebarListItem,
+  SidebarListItemAnchorLink,
+  SidebarListItemLink
+};
