+++ src/partials/sidebar/styles/sidebar-list-item-anchor-link.js
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+import SidebarListItemLinkStyle from "./sidebar-list-item-link-style";
+
+const SidebarListItemAnchorLink = styled.a`
+  ${SidebarListItemLinkStyle}
+`;
+
+export default SidebarListItemAnchorLink;
