+++ src/partials/sidebar/styles/sidebar-list-item-link.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+import { NavLink } from "react-router-dom";
+import SidebarListItemLinkStyle from "./sidebar-list-item-link-style";
+
+const SidebarListItemLink = styled(NavLink)`
+  ${SidebarListItemLinkStyle}
+`;
+
+export default SidebarListItemLink;
