+++ src/partials/sidebar/styles/sidebar-list-item.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+const SidebarListItem = styled.li`
+  padding: 0;
+  margin: 0;
+  width: 100%;
+`;
+
+export default SidebarListItem;
